pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Joey Dreijer <jdreijer@schubergphilis.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.1"
typer = "^0.6.1"
PyYAML = "^6.0"
neomodel = "^4.0.8"
pydantic = "^1.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
